jd01_10 annotations
