--- src/by/it/akhmelev/jd01_10/matlab_prototype/interfaces/IAdd.java
@@ -1,7 +1,0 @@
-package by.it.akhmelev.jd01_10.matlab_prototype.interfaces;
-
-import by.it.akhmelev.jd01_10.matlab_prototype.vars.Var;
-
-public interface IAdd {
-    Var add(Var var);   //сложение
-}
