--- src/by/it/akhmelev/jd01_10/matlab_prototype/interfaces/IDiv.java
@@ -1,7 +1,0 @@
-package by.it.akhmelev.jd01_10.matlab_prototype.interfaces;
-
-import by.it.akhmelev.jd01_10.matlab_prototype.vars.Var;
-
-public interface IDiv {
-    Var div(Var var);   //деление
-}
