--- src/by/it/akhmelev/jd01_10/matlab_prototype/interfaces/IMul.java
@@ -1,7 +1,0 @@
-package by.it.akhmelev.jd01_10.matlab_prototype.interfaces;
-
-import by.it.akhmelev.jd01_10.matlab_prototype.vars.Var;
-
-public interface IMul {
-    Var mul(Var var);   //умножение
-}
