+++ src/by/it/akhmelev/jd01_10/matlab_v2/Error.java
@@ -1,4 +1,4 @@
-package by.it.akhmelev.jd01_10.matlab_prototype;
+package by.it.akhmelev.jd01_10.matlab_v2;
 
 public class Error {
     //должен наследоваться extends Exception
