+++ src/by/it/akhmelev/jd01_10/matlab_v2/Patterns.java
@@ -1,4 +1,4 @@
-package by.it.akhmelev.jd01_10.matlab_prototype;
+package by.it.akhmelev.jd01_10.matlab_v2;
 
 public class Patterns {
     //регулярки
