+++ src/by/it/akhmelev/jd01_10/matlab_v2/interfaces/IMul.java
@@ -1,0 +1,7 @@
+package by.it.akhmelev.jd01_10.matlab_v2.interfaces;
+
+import by.it.akhmelev.jd01_10.matlab_v2.vars.Var;
+
+public interface IMul {
+    Var mul(Var var);   //умножение
+}
