+++ src/by/it/akhmelev/jd01_10/matlab_v2/interfaces/IVar.java
@@ -1,4 +1,4 @@
-package by.it.akhmelev.jd01_10.matlab_prototype.interfaces;
+package by.it.akhmelev.jd01_10.matlab_v2.interfaces;
 
 public interface IVar extends IAdd,IDiv,IMul,ISub {
     String toString();            //метод вывода переменной в строку
