+++ src/by/it/akhmelev/jd01_10/matlab_v2/vars/VarM.java
@@ -1,4 +1,4 @@
-package by.it.akhmelev.jd01_10.matlab_prototype.vars;
+package by.it.akhmelev.jd01_10.matlab_v2.vars;
 
 public class VarM extends Var {
     //это просто пример ;)
